pyproject.toml
[project]
name = "langchain-course"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "black>=25.12.0",
    "isort>=7.0.0",
    "langchain>=1.1.3",
    "langchain-openai>=1.1.2",
    "python-dotenv>=1.2.1",
]
